Data

+++ app/src/main/java/com/example/firebase/data/model/MainTraning.kt
@@ -7,7 +7,7 @@ data class MainTraning (
     @PrimaryKey(autoGenerate = true)
     val id : Int = 0 ,
     val  point : LatlngPoints,
-    val distance : Int,
+    val distance : Double,
     val duration : Long,
     val startAt: Long,
     val finishAt: Long,
